extractive-summarization/deploy/requirements.txt
pytorch-lightning==1.3.3
torchmetrics==0.5
